requirements-dev.txt
flake8
pylint
django-debug-toolbar
